+++ apps/fluent-tester/src/TestComponents/Icon/IconTest.tsx
@@ -6,8 +6,9 @@ import { Icon, RasterImageIconProps, SvgIconProps, FontIconProps } from '@fluent
 import { Test, TestSection, PlatformStatus } from '../Test';
 import { ICON_TESTPAGE } from './consts';
 import { E2ETestingIcon } from './IconE2ETest';
+import { IconExperimental } from './IconExperimental';
 
-const testImage = require('../../FluentTester/test-data/icon_24x24.png');
+const testImage = require('../../../../assets/icon_24x24.png');
 const testTtf = require('./../../../assets/Font_Awesome_900.otf');
 import TestSvg from './assets/test.svg';
 
@@ -103,6 +104,10 @@ const iconSections: TestSection[] = [
   {
     name: 'Icon for E2E Testing',
     component: E2ETestingIcon,
+  },
+  {
+    name: 'Default Icon',
+    component: IconExperimental,
   },
 ];