+++ packages/experimental/Icon/src/index.ts
@@ -1,2 +1,3 @@
 export * from './Icon.types';
 export * from './Icon';
+export { createIconProps } from './useIconProps.hooks';