+++ packages/components/Callout/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @fluentui-react-native/callout
 
-This log was last generated on Fri, 11 Nov 2022 18:06:32 GMT and should not be manually modified.
+This log was last generated on Wed, 16 Nov 2022 08:47:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.21.36
+
+Wed, 16 Nov 2022 08:47:25 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.21.1
+- Bump @uifabricshared/foundation-compose to v1.12.21
+
+## 0.21.35
+
+Sun, 13 Nov 2022 09:17:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.21.0
 
 ## 0.21.34