+++ packages/components/MenuButton/CHANGELOG.md
@@ -1,8 +1,30 @@
 # Change Log - @fluentui-react-native/menu-button
 
-This log was last generated on Fri, 11 Nov 2022 18:06:31 GMT and should not be manually modified.
+This log was last generated on Wed, 16 Nov 2022 08:47:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.1
+
+Wed, 16 Nov 2022 08:47:25 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.31.1
+- Bump @fluentui-react-native/contextual-menu to v0.20.1
+- Bump @fluentui-react-native/icon to v0.15.1
+- Bump @uifabricshared/foundation-compose to v1.12.21
+
+## 0.9.0
+
+Sun, 13 Nov 2022 09:17:06 GMT
+
+### Minor changes
+
+- fix double clicking on inline links (email not defined)
+- Bump @fluentui-react-native/button to v0.31.0
+- Bump @fluentui-react-native/contextual-menu to v0.20.0
+- Bump @fluentui-react-native/icon to v0.15.0
 
 ## 0.8.67