+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,28 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Fri, 11 Nov 2022 18:06:32 GMT and should not be manually modified.
+This log was last generated on Wed, 16 Nov 2022 08:47:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.15.1
+
+Wed, 16 Nov 2022 08:47:25 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.16.12
+- Bump @fluentui-react-native/framework to v0.8.20
+- Bump @fluentui-react-native/text to v0.16.1
+- Bump @fluentui-react-native/theming-utils to v0.18.3
+
+## 0.15.0
+
+Sun, 13 Nov 2022 09:17:06 GMT
+
+### Minor changes
+
+- fix double clicking on inline links (email not defined)
+- Bump @fluentui-react-native/text to v0.16.0
 
 ## 0.14.1